requirements.txt
asn1crypto==1.0.1
astroid==2.3.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.6
django-easy-pdf==0.2.0.dev1
django-pandas==0.6.1
django-paypal==1.0.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
html5lib==1.0.1
idna==2.8
isort==4.3.21
joblib==0.14.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.0
pycparser==2.19
pylint==2.4.2
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.0
urllib3==1.25.7
WeasyPrint==50
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
